requirements.txt
opencv-python>=4.5.0  # For video processing and image manipulation
scipy>=1.7.0      # For image processing and filtering
torch>=2.0.0      # For GPU-accelerated image processing
torchvision>=0.15.0  # For image transformations
numpy>=1.21.0     # For numerical operations
Pillow>=9.0.0     # For image processing
tqdm>=4.65.0      # For progress bars
matplotlib>=3.5.0  # For visualization
seaborn>=0.11.0    # For advanced visualization
pandas>=1.4.0      # For data handling
scikit-learn>=1.0.0  # For metrics and validation

# Added for synthetic data generation
diffusers>=0.20.0
transformers>=4.30.0
accelerate>=0.20.0

# Added for ChatList-KI task management
sqlalchemy>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0     # For running tests
pytest-cov>=3.0.0  # For test coverage

# Added for CI/CD pipeline
tabulate>=0.8.10   # For reporting
pyserial>=3.5      # For performance log analysis
fpdf>=1.7.2        # For report generation
markdown>=3.3.0    # For documentation
jq>=1.2.0          # For JSON processing
onnx>=1.12.0       # For model optimization
tensorflow>=2.9.0  # For model optimization